requirements.txt
# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
camelot-py[cv]==0.11.0

# Text Processing
spacy==3.7.2
nltk==3.8.1
langchain==0.0.350
tiktoken==0.5.2

# Embeddings & Vector Store
openai==1.3.5
httpx==0.25.2  # Specific version for compatibility
pinecone-client==2.2.4
chromadb==0.4.18

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
tqdm==4.66.1
pandas==2.1.3
numpy==1.26.2